# ENH: SampleSurfaceMesh and GeometryMath Speed Optimizations