Fixes #962
Fixes #961
Updated Documentation

| Version | Time | CPU | Iterations |
|---------|------|-----|------------|
| Baseline | 7232 ns | 7231 ns | 75982 |
| Post-Cache | 6249 ns | 6249 ns | 107611 |
| branch-reduction | 6234 ns | 6234 ns | 104425 |

Todo:

- [x] add caching routing
- [x] strip extraneous copies
- [x] extract all major tight loop constructions
- [x] extract extraneous if statements
- [x] consolidate calculations for coditionals in tight loops
- [x] Strip `if` from `DoesRayIntersectBox`
- [x] Strip `if` from `IsPointInTriangle`
- [x] Strip `if` from `RayCrossesTriangle`
- [x] `Ray.hpp` now uses a cached variant in the algorithm to avoid inserting Eigen (after discussion) while preserving API