@JDuffeyBQ 2024-07-22
Do you have the code that you used for the timings?