@nyoungbq 2024-07-22
> Do you have the code that you used for the timings?

I slacked it to you because it uses hardcoding to get around additional library linkage